third_party/blink/web_tests/fast/js/mozilla/strict/15.10.7-expected.txt

PASS 'use strict'; var r = /foo/; r.source = "bar"; r.source threw exception of type TypeError.
PASS var r = /foo/; r.source = "bar"; r.source is "foo"
PASS true === true
FAIL 'use strict'; var r = /foo/; delete r.source should throw an instance of TypeError
FAIL var r = /foo/; delete r.source should be false. Was true.
PASS true === true
PASS 'use strict'; var r = /foo/; r.global = true; r.global threw exception of type TypeError.
PASS var r = /foo/; r.global = true; r.global is false
PASS true === true
FAIL 'use strict'; var r = /foo/; delete r.global should throw an instance of TypeError
FAIL var r = /foo/; delete r.global should be false. Was true.
PASS true === true
PASS 'use strict'; var r = /foo/; r.ignoreCase = true; r.ignoreCase threw exception of type TypeError.
PASS var r = /foo/; r.ignoreCase = true; r.ignoreCase is false
PASS true === true
FAIL 'use strict'; var r = /foo/; delete r.ignoreCase should throw an instance of TypeError
FAIL var r = /foo/; delete r.ignoreCase should be false. Was true.
PASS true === true
PASS 'use strict'; var r = /foo/; r.multiline = true; r.multiline threw exception of type TypeError.
PASS var r = /foo/; r.multiline = true; r.multiline is false
PASS true === true
FAIL 'use strict'; var r = /foo/; delete r.multiline should throw an instance of TypeError
FAIL var r = /foo/; delete r.multiline should be false. Was true.
PASS true === true
PASS 'use strict'; var r = /foo/; r.lastIndex = 42; r.lastIndex is 42
PASS var r = /foo/; r.lastIndex = 42; r.lastIndex is 42
PASS true === true
PASS 'use strict'; var r = /foo/; delete r.lastIndex threw exception of type TypeError.
PASS var r = /foo/; delete r.lastIndex is false
PASS true === true
 PASSED! 
PASS successfullyParsed is true

TEST COMPLETE
